move some evm helpers
