+++ modules/evm/src/lib.rs
@@ -58,10 +58,9 @@ pub use module_support::{
 pub use orml_traits::{currency::TransferAll, MultiCurrency};
 use primitive_types::{H160, H256, U256};
 pub use primitives::{
-	convert_decimals_from_evm, convert_decimals_to_evm,
 	evm::{
-		CallInfo, CreateInfo, EvmAddress, ExecutionInfo, Vicinity, MIRRORED_NFT_ADDRESS_START,
-		MIRRORED_TOKENS_ADDRESS_START, PREDEPLOY_ADDRESS_START,
+		convert_decimals_from_evm, convert_decimals_to_evm, CallInfo, CreateInfo, EvmAddress, ExecutionInfo, Vicinity,
+		MIRRORED_NFT_ADDRESS_START, MIRRORED_TOKENS_ADDRESS_START, PREDEPLOY_ADDRESS_START,
 	},
 	task::TaskResult,
 	Balance, CurrencyId, ReserveIdentifier,
