+++ modules/evm/src/runner/stack.rs
@@ -39,8 +39,7 @@ use module_evm_utility::{
 };
 use module_support::AddressMapping;
 pub use primitives::{
-	convert_decimals_from_evm,
-	evm::{EvmAddress, Vicinity, MIRRORED_NFT_ADDRESS_START},
+	evm::{convert_decimals_from_evm, EvmAddress, Vicinity, MIRRORED_NFT_ADDRESS_START},
 	ReserveIdentifier,
 };
 use sha3::{Digest, Keccak256};
